refactor: decompose ActivityFeedPanel into smaller modules
- Break down ActivityFeedComponents.tsx into individual component files
- Create a better organized directory structure for activity feed components
- Extract UI rendering logic into ActivityFeedContent component
- Move stateful logic to custom useActivityFeedLayout hook
- Update tests to work with the new component structure
- All tests passing with improved code organization
