+++ src/components/dashboard/activityFeed/index.ts
@@ -1,0 +1,20 @@
+// Re-export components
+export { ActivityFeedHeader } from "./components/ActivityFeedHeader";
+export { ActivityFeedLoading } from "./components/ActivityFeedLoading";
+export { ActivityFeedError } from "./components/ActivityFeedError";
+export { ActivityFeedEmpty } from "./components/ActivityFeedEmpty";
+export { IncrementalLoadingIndicator } from "./components/IncrementalLoadingIndicator";
+export { CommitItem } from "./components/CommitItem";
+export { CommitList } from "./components/CommitList";
+export { LoadMoreSection } from "./components/LoadMoreSection";
+export { ActivityFeedContent } from "./components/ActivityFeedContent";
+
+// Re-export hooks
+export { useActivityFeedLayout } from "./hooks/useActivityFeedLayout";
+
+// Re-export utility functions
+export {
+  formatCommitDate,
+  calculateListHeight,
+  extractCommitTitle,
+} from "./utils/activityFeedUtils";
